Aadded-server-side-authorization

+++ backend/routes/posts.js
@@ -29,13 +29,15 @@ const router = express.Router();
 
 router.post(
   "",
+  checkhAuth,
   multer({storage: storage}).single('image'),
   (req, res, next) => {
     const url = req.protocol + '://' + req.get('host');
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + '/images/' + req.file.filename
+      imagePath: url + '/images/' + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -51,24 +53,34 @@ router.post(
 
 });
 
-router.put("/:id", multer({storage: storage}).single('image'), (req, res, next) => {
-  let imagePath = req.body.imagePath;
-  if (req.file) {
-    const url = req.protocol + '://' + req.get('host');
-    imagePath = url + '/images/' + req.file.filename ;
+router.put(
+  "/:id",
+  checkhAuth,
+  multer({ storage: storage }).single("image"),
+  (req, res, next) => {
+    let imagePath = req.body.imagePath;
+    if (req.file) {
+      const url = req.protocol + "://" + req.get("host");
+      imagePath = url + "/images/" + req.file.filename;
+    }
+    const post = new Post({
+      _id: req.body.id,
+      title: req.body.title,
+      content: req.body.content,
+      imagePath: imagePath
+    });
+    Post.updateOne(
+      { _id: req.params.id, creator: req.userData.userId },
+      post
+    ).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: "Update successful!" });
+      } else {
+        res.status(401).json({ message: "Not authorized!" });
+      }
+    });
   }
-  const post = new Post({
-    _id: req.body.id,
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: imagePath
-  });
-  console.log(post);
-  Post.updateOne({_id: req.params.id},post).then(result => {
-    console.log(result);
-    res.status(200).json({message: 'Update successful!'});
-  });
-});
+);
 
 router.get("", (req, res, next) => {
   const pageSize = +req.query.pagesize;
@@ -105,9 +117,12 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkhAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({message: 'Deletion successful!'});
+    } else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
   });
 });
